+++ src/main/kotlin/lotto/Application.kt
@@ -1,5 +1,9 @@
 package lotto
 
+import lotto.domain.service.LottoGenerator
+import lotto.domain.service.LottoRankFinder
+import lotto.view.LottoView
+
 fun main() {
     // 의존성 주입
     val lottoView = LottoView()
