+++ src/main/kotlin/lotto/domain/service/LottoRankFinder.kt
@@ -1,4 +1,4 @@
-package lotto
+package lotto.domain.service
 
 import lotto.domain.Lotto
 import lotto.domain.Rank
