+++ src/test/kotlin/lottoTest/domain/ScoreRankMapTest.kt
@@ -1,8 +1,8 @@
 package lottoTest.domain
 
-import lotto.LottoRankFinder
 import lotto.domain.Lotto
 import lotto.domain.LottoNumber
+import lotto.domain.service.LottoRankFinder
 import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.DisplayName
 import org.junit.jupiter.api.Test
